# openapi.yaml
openapi: 3.1.0
info:
  title: PIM/DAM API — Beverage Catalog
  version: 1.0.0
  description: >
    Product Information & Digital Asset Management with a dynamic attribute
    model. Authenticate via POST /auth/login, then send the access token as
    `Authorization: Bearer <token>` (1h TTL; rotate with POST /auth/refresh).
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  schemas:
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        sku: { type: string }
        name: { type: string }
        status: { type: string, enum: [draft, in_review, published, archived] }
        version: { type: integer, description: Optimistic-locking version; echo on PATCH }
        attributes:
          type: object
          additionalProperties: true
          description: Dynamic values keyed by attribute code, validated per product type
...
paths:
...
  /products:
    get:
      summary: List/search products (needs product.view)
      parameters:
        - { name: q, in: query, schema: { type: string }, description: Full-text + fuzzy search }
        - { name: type, in: query, schema: { type: string }, description: Product type code }
        - { name: status, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      description: "Also supports attr.<code>=<value> filters via JSONB containment (e.g. attr.beer_style=IPA)."
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Product" } }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
    post:
      summary: Create product (needs product.edit); attributes validated against type
      responses:
        "201": { description: Created }
        "422": { description: Attribute validation errors }
  /products/{id}:
    get:
      summary: Full product with variants, assets, attribute_schema
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Product }, "404": { description: Not found } }
    patch:
      summary: Partial update with optimistic locking (version required)
      responses:
        "200": { description: Updated (version incremented) }
        "403": { description: Publishing requires product.publish }
        "409": { description: Version conflict — reload and retry }
        "422": { description: Attribute validation errors }
...
  /assets/file/{key}:
    get:
      summary: Serve original or rendition file (immutable cache headers)
      parameters: [{ name: key, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: File bytes }, "404": { description: Not found } }
...